requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.12
Django==4.1.5
django-allauth==0.48.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.3
platformdirs==2.6.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.14
